Moving Menu Items to Settings Menu
