+++ public/locales/en-US/main.json
@@ -11,7 +11,7 @@
   "export": "Export",
   "author": "Made by Jing Hua",
   "about": "About",
-  "api": "API",
+  "apiSettings": "API Settings",
   "personal": "Personal",
   "free": "Free",
   "downloadChat": "Download Chat",
