+++ public/locales/en/model.json
@@ -22,7 +22,7 @@
     "label": "Frequency Penalty",
     "description": "Number between -2.0 and 2.0. Positive values penalise new tokens based on their existing frequency in the text so far, decreasing the model's likelihood to repeat the same line verbatim. (Default: 0)"
   },
-  "defaultChatConfig": "Default Chat Config",
+  "defaultChatConfig": "Model Settings for New Chats",
   "defaultSystemMessage": "Default System Message",
   "resetToDefault": "Reset To Default"
 }
