+++ public/locales/ru/main.json
@@ -11,7 +11,7 @@
   "export": "Экспорт",
   "author": "Автор: Jing Hua",
   "about": "О программе",
-  "api": "API",
+  "apiSettings": "Настройки API",
   "personal": "Личный",
   "free": "Бесплатный",
   "downloadChat": "Скачать чат",
