+++ src/components/Menu/MenuOptions/MenuOptions.tsx
@@ -1,13 +1,9 @@
-import React from 'react';
 import useStore from '@store/store';
 
-import Api from './Api';
-import Me from './Me';
 import AboutMenu from '@components/AboutMenu';
 import ImportExportChat from '@components/ImportExportChat';
 import SettingsMenu from '@components/SettingsMenu';
 import CollapseOptions from './CollapseOptions';
-import GoogleSync from '@components/GoogleSync';
 import TokensCostMenu from './TokensCostMenu'
 import AuthenticatedUserLogout from './AuthenticatedUserLogout';
 
@@ -25,11 +21,7 @@ const MenuOptions = () => {
         } overflow-hidden transition-all`}
       >
         {countTotalTokens && <TokensCostMenu />}
-        {/* {googleClientId && <GoogleSync clientId={googleClientId} />} */}
-        <ImportExportChat />
         <SettingsMenu />
-        <Api />
-        {/* <Me /> */}
         <AboutMenu />
         <AuthenticatedUserLogout />
       </div>
